+++ src/api-posix.c
@@ -61,3 +61,7 @@ SEXP psll_terminate(SEXP p) {
 SEXP psll_kill(SEXP p) {
   return psll_send_signal(p, ScalarInteger(SIGKILL));
 }
+
+SEXP psll_interrupt(SEXP p, SEXP ctrlc, SEXP interrupt_path) {
+  return psll_send_signal(p, ScalarInteger(SIGINT));
+}