+++ NEWS.md
@@ -5,6 +5,9 @@
 
 * `ps_open_files()` lists all open files of a process.
 
+* `ps_interrupt()` interrupts a process. It sends a `SIGINT` signal on
+  POSIX systems, and it can send CTRL+C or CTRL+BREAK events on Windows.
+
 # ps 1.0.0
 
 First released version.